pyproject.toml
[project]
name = "evo2-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "umap-learn>=0.5.7",
]
